Add confirm message to delete product

+++ public/javascripts/main.js
@@ -1,16 +1,18 @@
 $(document).ready(function(){
   $('.deleteProduct').on('click', function(e){
-    // debugger
     pid = $(this).data('id')
     url = '/products/delete/' + pid;
-    // console.log(url);
-    $.ajax({
-      url: url,
-      type: 'DELETE',
-      success: function(result){
-        console.log('deleted success!')
-      }
-    });
-    window.location = '/';
+
+    var confirmed = confirm("Are you sure?");
+    if(confirmed){
+      $.ajax({
+        url: url,
+        type: 'DELETE',
+        success: function(result){
+          console.log('deleted success!')
+        }
+      });
+      window.location = '/';
+    }
   });
 });
